[Chore] PostgreSQL 도입 - 의존성 추가 - Id (Short-> Long) 수정

+++ src/main/java/com/sparta/moit/domain/meeting/dto/RegionFirstResponseDto.java
@@ -6,10 +6,10 @@ import lombok.Getter;
 
 @Getter
 public class RegionFirstResponseDto {
-    private Short regionFirstId;
+    private Long regionFirstId;
     private String regionFirstName;
     @Builder
-    public RegionFirstResponseDto(Short regionFirstId, String regionFirstName) {
+    public RegionFirstResponseDto(Long regionFirstId, String regionFirstName) {
         this.regionFirstId = regionFirstId;
         this.regionFirstName = regionFirstName;
     }
